feat(dotfiles)!: update

+++ .files/.config/fish/conf.d/aliases.fish
@@ -6,17 +6,20 @@ if not status is-interactive
 end
 
 # Useful general shortcuts
-abbr c clear
+abbr c ' clear'
+abbr clear ' clear'
 abbr cal 'task cal; or ncal -wC3; or command cal -B1 -A1'
-alias start '{ test -f .jig* && jig start .; } ||
-	{ jig ls | fzf | xargs jig start; }'
 
 alias update 'brew update && brew outdated'
 alias upgrade 'ya pack -u && brew upgrade'
 alias outdated 'brew outdated'
+alias work 'cd ~/code/work'
 alias wiki 'cd ~/code/rafi/*/content/wiki'
+alias start '{ test -f .jig* && jig start .; } ||
+	{ jig ls | fzf | xargs jig start; }'
 
-abbr -a --position anywhere --set-cursor -- -h "-h 2>&1 | bat --plain --language=help"
+abbr -a --position anywhere --function last_history_item -- !!
+abbr -a --position anywhere --set-cursor -- -h "-h 2>&1 | bat -pl=help"
 
 if [ $OS_DARWIN ]
 	command -q gdircolors; and abbr dircolors gdircolors
@@ -55,7 +58,7 @@ if command -q nvim
 	abbr vim nvim
 	abbr nvi nvim
 	abbr suvim sudo -E nvim
-	# alias vimpager 'nvim - -c "lua require(\'util\').colorize()"'
+	abbr vimpager 'nvim - -c "lua Snacks.terminal.colorize()"'
 	# alias vless="nvim -u $PREFIX/share/nvim/runtime/macros/less.vim"
 else
 	abbr v 'vim (fzf)'
@@ -112,24 +115,28 @@ end
 
 # }}}
 # systemctl ------------------------------------------ sc for systemctl -- {{{
-abbr sc systemctl
-abbr scu "systemctl --user"
-abbr scs "command systemctl status"
-abbr scl "systemctl --type service --state running"
-abbr sclu "systemctl --user --type service --state running"
-abbr sce "sudo systemctl enable --now"
-abbr scd "sudo systemctl disable --now"
-abbr scr "sudo systemctl restart"
-abbr sco "sudo systemctl stop"
-abbr sca "sudo systemctl start"
-abbr scf "systemctl --failed --all"
+if [ $OS_LINUX ]
+	abbr sc systemctl
+	abbr scu "systemctl --user"
+	abbr scs "command systemctl status"
+	abbr scl "systemctl --type service --state running"
+	abbr sclu "systemctl --user --type service --state running"
+	abbr sce "sudo systemctl enable --now"
+	abbr scd "sudo systemctl disable --now"
+	abbr scr "sudo systemctl restart"
+	abbr sco "sudo systemctl stop"
+	abbr sca "sudo systemctl start"
+	abbr scf "systemctl --failed --all"
+end
 
 # }}}
 # journalctl ---------------------------------------- j+ for journalctl -- {{{
-abbr jb "journalctl -b"
-abbr jf "journalctl --follow"
-abbr jg "journalctl -b --grep"
-abbr ju "journalctl --unit"
+if [ $OS_LINUX ]
+	abbr jb "journalctl -b"
+	abbr jf "journalctl --follow"
+	abbr jg "journalctl -b --grep"
+	abbr ju "journalctl --unit"
+end
 
 # }}}
 # Git ------------------------------------------------------- g for git -- {{{
@@ -152,6 +159,9 @@ abbr gs  git status -sb
 abbr gl  git lg -15
 abbr gll git lg
 abbr gld git lgd -15
+
+# Promoted git aliases
+abbr ck git checkout
 
 # }}}
 # Docker ------------------------------------------------- d for docker -- {{{
@@ -189,6 +199,12 @@ abbr kgpy kubectl get pod -o yaml
 abbr krr kubectl rollout restart deploy
 
 # }}}
+# Virtualization / Emulation --------------------------------------------- {{{
+# See: https://github.com/lima-vm/lima
+abbr li  lima
+abbr lic limactl
+
+# }}}
 # Processes -------------------------------------------------------------- {{{
 abbr process ps -ax
 abbr pst pstree -g 3 -ws
@@ -198,11 +214,6 @@ abbr pst pstree -g 3 -ws
 # See: https://github.com/casey/just
 abbr j just
 abbr jc just --choose
-
-# }}}
-# Machines --------------------------------------------- m for machines -- {{{
-# See: https://github.com/lima-vm/lima
-abbr m limactl
 
 # }}}
 # XDG conformity --------------------------------------------------------- {{{
